Staging site: rename file useDeleteStagingSite
