+++ client/my-sites/hosting/staging-site-card/use-delete-staging-site.ts
